Changes to structure of database module
