+++ sayn/database/postgresql.py
@@ -3,7 +3,7 @@ import io
 from sqlalchemy import create_engine
 
 from ..core.errors import Ok
-from .database import Database
+from . import Database
 
 db_parameters = ["host", "user", "password", "port", "dbname"]
